% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify.R
\name{spotify_chart}
\alias{spotify_chart}
\title{Scrapes spotify chart data}
\usage{
spotify_chart(start = "2018/10/26", end = Sys.Date(),
  url = "https://spotifycharts.com/regional/gb/weekly/")
}
\arguments{
\item{start}{Defaults to "2019/01/03"}

\item{end}{defaults to "2019/12/19"}

\item{url}{defaults to "https://spotifycharts.com/regional/gb/weekly/latest/"}
}
\description{
Scrapes spotify chart data
}
\examples{
## d<-spotify_chart()



}
